Instrument initial setup of a new slave and better handling for promotion
This changes the data directory initialization to use pg_basebackup to load the
data directory from the master when provisioning a new slave. With this,
provisioning a new slave is done end to end with a single run, and no need to
manually pull over the data directory.

Also adjusts some configuration settings that can be set on both the master and
slaves. Keeping these in sync between the two makes the slave better prepped if
it gets promoted to the master. These include WAL related settings that will
allow it have slaves without needing its configuraion updated and server
restarted.